requirements.txt
mss
pillow
pytesseract
tkinter
